feat!: Implement heartbeat mechanism
The server will shutdown when no client is connected to the server for a
period of time.

BREAKING CHANGE: No longer tracks the lifespan of the wrapper command
